changed index page and layout
